requirements.txt
Django>=5.0
mysqlclient
djangorestframework
requests
pytest
pytest-django